Conflicts:
	CHANGELOG
	cocos2d-ios.xcodeproj/project.pbxproj
	cocos2d-mac.xcodeproj/project.pbxproj
	cocos2d/CCGrid.m
	cocos2d/CCRenderTexture.h
	cocos2d/CCRenderTexture.m
	cocos2d/CCTextureAtlas.m
	cocos2d/ccMacros.h
	tests/ActionsTest.m
	tests/MenuTest.m
	tests/RenderTextureTest.m
	tests/TileMapTest.m

Former-commit-id: 231e5cfa774028731b8e02235ba1dc5f8bb72b2f

+++ cocos2d/CCSpriteBatchNode.h
@@ -115,6 +115,6 @@
 -(void) removeSpriteFromAtlas:(CCSprite*)sprite;
 
 -(NSUInteger) rebuildIndexInOrder:(CCSprite*)parent atlasIndex:(NSUInteger)index;
--(NSUInteger) atlasIndexForChild:(CCSprite*)sprite atZ:(int)z;
+-(NSUInteger) atlasIndexForChild:(CCSprite*)sprite atZ:(NSInteger)z;
 
 @end
